requirements-dev.txt
autoflake
autopep8
factory-boy
flake8
isort
pytest
pytest-cov
pytest-django
